+++ R/wikidata.R
@@ -49,7 +49,7 @@ dbpedia_get_wikidata_uris <- function(x, optional, endpoint, limit = 100, wait =
   if (verbose) cli_alert_info("input length: {.val {length(x)}}")
   x <- unique_msg(x, verbose = verbose)
   
-  template <- 'SELECT distinct ?item ?wikidata_uri ?key
+  template <- 'SELECT distinct ?item ?wikidata_uri
       WHERE {
       VALUES ?item {%s}
       ?item owl:sameAs ?wikidata_uri
@@ -112,11 +112,12 @@ dbpedia_get_wikidata_uris <- function(x, optional, endpoint, limit = 100, wait =
 #' @rdname wikidata_query
 setGeneric("wikidata_query", function(x, ...) standardGeneric("wikidata_query"))
 
+#' @param language Language to be used in Wikidata query.
 #' @rdname wikidata_query
 #' @examples
 #' wikidata_ids <- c("Q1741365", "Q3840", "Q437")
 #' wikidata_query(wikidata_ids, id = "P439",progress = TRUE)
-setMethod("wikidata_query", "character", function(x, id, limit = 100L, wait = 1, verbose = TRUE, progress = FALSE){
+setMethod("wikidata_query", "character", function(x, id, language = getOption("dbpedia.lang"), limit = 100L, wait = 1, verbose = TRUE, progress = FALSE){
   
   if (!requireNamespace("WikidataQueryServiceR", quietly = TRUE)){
     stop("R package WikidataQueryServiceR required but not available. ")
@@ -139,8 +140,8 @@ setMethod("wikidata_query", "character", function(x, id, limit = 100L, wait = 1,
         VALUES ?item { %s }
         OPTIONAL { ?item wdt:%s ?key . }
         ?item rdfs:label ?label
-          filter(lang(?label) = "de")
-        SERVICE wikibase:label { bd:serviceParam wikibase:language "de". }
+          filter(lang(?label) = "%s")
+        SERVICE wikibase:label { bd:serviceParam wikibase:language "%s". }
       }'
   
   chunks <- as_chunks(x = x, size = limit)
@@ -158,7 +159,9 @@ setMethod("wikidata_query", "character", function(x, id, limit = 100L, wait = 1,
     query <- sprintf(
       template,
       paste0("wd:", chunks[[i]], collapse = " "),
-      id
+      id,
+      language,
+      language
     )
     
     if (i > 1L) Sys.sleep(wait)
@@ -187,7 +190,7 @@ setMethod("wikidata_query", "character", function(x, id, limit = 100L, wait = 1,
 setMethod(
   "wikidata_query",
   "data.table",
-  function(x, id, limit = 100L, wait = 1, verbose = TRUE, progress = FALSE){
+  function(x, id, language = getOption("dbpedia.lang"), limit = 100L, wait = 1, verbose = TRUE, progress = FALSE){
     if (!"wikidata_id" %in% colnames(x)){
       cli_alert_danger("{.fn wikidata_query} requires column {.val wikidata_id}")
       stop()
@@ -202,9 +205,16 @@ setMethod(
       progress = progress
     )
     dt <- as.data.table(tbl, key = "wikidata_uri")
+    dt[, (paste(id, "key", sep = "_")) := NULL]
+    dt_min <- dt[,{
+      list(id = paste(.SD[[paste(id, "keyLabel", sep = "_")]], collapse = "|"))
+      }, by = c("wikidata_uri", "wikidata_label")
+    ]
+    setnames(dt_min, old = "id", new = id)
+    setkeyv(x = dt_min, cols = "wikidata_uri")
     setkeyv(x = x, cols = "wikidata_uri")
-    y <- dt[x]
-    setcolorder(x = y, neworder = c(colnames(x), colnames(dt)[-which(colnames(dt) == "wikidata_uri")]))
+    y <- dt_min[x]
+    setcolorder(x = y, neworder = c(colnames(x), colnames(dt_min)[-which(colnames(dt_min) == "wikidata_uri")]))
     y
   }
 )
@@ -221,30 +231,24 @@ setGeneric(
 
 
 #' @examples
-#' # This example is commented out because we get an error that the SSL 
-#' # certificate has expired
+#' library(dbpedia)
+#' library(quanteda)
+#' library(dplyr)
+#' library(tibble)
 #' 
-#' # options(dbpedia.lang = "en")
-#' # options(dbpedia.endpoint = "http://api.dbpedia-spotlight.org/en/annotate")
-#' # library(quanteda)
+#' options(dbpedia.lang = "en")
+#' options(dbpedia.endpoint = "http://api.dbpedia-spotlight.org/en/annotate")
 #' 
-#' # dt <- data_char_ukimmig2010 |>
-#' #   corpus() |>
-#' #  get_dbpedia_uris(verbose = FALSE) %>% 
-#' #   add_wikidata_uris(endpoint = "http://de.dbpedia.org/sparql", progress = TRUE)
-#' 
-#' library(polmineR)
-#' use("GermaParl2")
-#' 
-#' options(dbpedia.lang = "de")
-#' options(dbpedia.endpoint = "http://localhost:2222/rest/annotate")
-#' 
-#' data <- polmineR::corpus("GERMAPARL2MINI") %>% 
-#'   subset(speaker_name == "Carlo Schmid") %>%
-#'   subset(p_type == "speech") %>% 
-#'   get_dbpedia_uris(language = "de", s_attribute = "ne", max_len = 5067) %>% 
-#'   add_wikidata_uris(endpoint = "http://de.dbpedia.org/sparql", verbose = FALSE, progress = TRUE) %>% 
-#'   wikidata_query(id = "P31")
+#' httr::set_config(httr::config(ssl_verifypeer = 0L))
+#'
+#' uritab <- data_char_ukimmig2010 |>
+#'   corpus() |>
+#'   get_dbpedia_uris() %>% 
+#'   add_wikidata_uris(endpoint = "https://dbpedia.org/sparql/", progress = TRUE, limit = 50) %>% 
+#'   wikidata_query(id = "P31") %>% 
+#'   as_tibble() %>% 
+#'   filter(grepl("political party", P31))
+#'   
 #' @rdname wikidata_uris
 #' @importFrom data.table setkeyv
 setMethod(